Make active_admin_comments migration idempotent to fix deployment duplicate table error

+++ db/migrate/20251223155428_create_active_admin_comments.rb
@@ -1,16 +1,16 @@
 class CreateActiveAdminComments < ActiveRecord::Migration[8.1]
   def self.up
-    create_table :active_admin_comments do |t|
+    create_table :active_admin_comments, if_not_exists: true do |t|
       t.string :namespace
       t.text   :body
       t.references :resource, polymorphic: true
       t.references :author, polymorphic: true
       t.timestamps
     end
-    add_index :active_admin_comments, [:namespace]
+    add_index :active_admin_comments, [:namespace], if_not_exists: true
   end
 
   def self.down
-    drop_table :active_admin_comments
+    drop_table :active_admin_comments if table_exists?(:active_admin_comments)
   end
 end
